refactor default data
