+++ public/index.html
@@ -37,8 +37,6 @@
       To create a production bundle, use `npm run build` or `yarn build`.
     -->
 </body>
-<script async defer src="https://apis.google.com/js/api.js" onload="this.onload=function(){};handleClientLoad()"
-  onreadystatechange="if (this.readyState === 'complete') this.onload()">
-</script>
+<script async defer src="https://apis.google.com/js/api.js" onload="window.handleClientLoad()"></script>
 
 </html>
