+++ src/YouTube/LoginPage.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { Component } from 'react'
 import styled from 'styled-components'
 import gapi from 'gapi-client'
 
@@ -10,47 +10,56 @@ const Grid = styled.section`
   grid-gap: 20px;
   padding: 20px;
 `
+// Options
+const CLIENT_ID =
+  '843342214316-febn2vufffq9heqdut8vhtlfvqjh15sd.apps.googleusercontent.com'
+const DISCOVERY_DOCS = [
+  'https://www.googleapis.com/discovery/v1/apis/youtube/v3/rest',
+]
+const SCOPES = 'https://www.googleapis.com/auth/youtube.readonly'
 
-export default function LoginPage() {
-  // Options
-  const CLIENT_ID =
-    '843342214316-febn2vufffq9heqdut8vhtlfvqjh15sd.apps.googleusercontent.com'
-  const DISCOVERY_DOCS = [
-    'https://www.googleapis.com/discovery/v1/apis/youtube/v3/rest',
-  ]
-  const SCOPES = 'https://www.googleapis.com/auth/youtube.readonly'
+const authorizeButton = document.getElementById('authorize-button')
+const signoutButton = document.getElementById('signout-button')
+const content = document.getElementById('content')
+const channelForm = document.getElementById('channel-form')
+const channelInput = document.getElementById('channel-input')
+const videoContainer = document.getElementById('video-container')
 
-  const authorizeButton = document.getElementById('authorize-button')
-  const signoutButton = document.getElementById('signout-button')
-  const content = document.getElementById('content')
-  const channelForm = document.getElementById('channel-form')
-  const channelInput = document.getElementById('channel-input')
-  const videoContainer = document.getElementById('video-container')
+const defaultChannel = 'techguyweb'
 
-  const defaultChannel = 'techguyweb'
+// const xhr = new XMLHttpRequest(),
+//   method = 'GET',
+//   url = 'https://apis.google.com/js/api.js'
 
-  // const xhr = new XMLHttpRequest(),
-  //   method = 'GET',
-  //   url = 'https://apis.google.com/js/api.js'
+// xhr.open(method, url, true)
+// xhr.onreadystatechange = function() {
+//   if (xhr.readyState === 'complete') {
+//     window.onload = this.handleClientLoad
+//     console.log('hello')
+//   }
+// }
 
-  // xhr.open(method, url, true)
-  // xhr.onreadystatechange = function() {
-  //   if (xhr.readyState === 'complete') {
-  //     window.onload = handleClientLoad
-  //     console.log('hello')
-  //   }
+export default class LoginPage extends Component {
+  constructor(props) {
+    super(props)
+    window.handleClientLoad = () => {
+      this.handleClientLoad()
+    }
+  }
+  // ComponentDidMount() {
+  //   loadjs('./js/main.js', function() {
+  //     loadjs('./js/common_scripts.js')
+  //   })
   // }
 
-  window.onload = handleClientLoad
-
   // Load auth2 library
-  function handleClientLoad() {
+  handleClientLoad() {
     console.log('handleClientLoad')
-    gapi.load('client:auth2', initClient)
+    gapi.load('client:auth2', this.initClient)
   }
 
   // Init API client library and set up sign in listeners
-  function initClient() {
+  initClient() {
     gapi.client
       .init({
         discoveryDocs: DISCOVERY_DOCS,
@@ -59,22 +68,22 @@ export default function LoginPage() {
       })
       .then(() => {
         // Listen for sign in state changes
-        gapi.auth2.getAuthInstance().isSignedIn.listen(updateSigninStatus)
+        gapi.auth2.getAuthInstance().isSignedIn.listen(this.updateSigninStatus)
         // Handle initial sign in state
-        updateSigninStatus(gapi.auth2.getAuthInstance().isSignedIn.get())
-        authorizeButton.onclick = handleAuthClick
-        signoutButton.onclick = handleSignoutClick
+        this.updateSigninStatus(gapi.auth2.getAuthInstance().isSignedIn.get())
+        authorizeButton.onclick = this.handleAuthClick
+        signoutButton.onclick = this.handleSignoutClick
       })
   }
 
   // Update UI sign in state changes
-  function updateSigninStatus(isSignedIn) {
+  updateSigninStatus(isSignedIn) {
     if (isSignedIn) {
       authorizeButton.style.display = 'none'
       signoutButton.style.display = 'block'
       content.style.display = 'block'
       videoContainer.style.display = 'block'
-      getChannel(defaultChannel)
+      this.getChannel(defaultChannel)
     } else {
       authorizeButton.style.display = 'block'
       signoutButton.style.display = 'none'
@@ -84,42 +93,44 @@ export default function LoginPage() {
   }
 
   // Handle login
-  function handleAuthClick() {
+  handleAuthClick() {
     gapi.auth2.getAuthInstance().signIn()
   }
 
   // Handle logout
-  function handleSignoutClick() {
+  handleSignoutClick() {
     gapi.auth2.getAuthInstance().signOut()
   }
 
   // Get channel from API
-  function getChannel(channel) {
+  getChannel(channel) {
     console.log(channel)
   }
 
-  return (
-    <Grid id="content">
-      Log In With Google
-      <button id="authorize-button">Log In</button>
-      <button id="signout-button">Log Out</button>
-      <div id="content">
-        <form id="channel-form">
-          <div id="input-field">
-            <input
-              id="channel-input"
-              placeholder="Enter Channel Name"
-              type="text"
-            />
-            <input value="Get Channel Data" type="submit" />
+  render() {
+    return (
+      <Grid id="content">
+        Log In With Google
+        <button id="authorize-button">Log In</button>
+        <button id="signout-button">Log Out</button>
+        <div id="content">
+          <form id="channel-form">
+            <div id="input-field">
+              <input
+                id="channel-input"
+                placeholder="Enter Channel Name"
+                type="text"
+              />
+              <input value="Get Channel Data" type="submit" />
 
-            {/* <input name="playlist" placeholder="Playlist name" type="text" />
-            <label htmlFor="playlist">Enter playlist name</label>
-            <input type="submit" value="Load videos from playlist" /> */}
-          </div>
-        </form>
-        <div id="video-container" />
-      </div>
-    </Grid>
-  )
+              {/* <input name="playlist" placeholder="Playlist name" type="text" />
+              <label htmlFor="playlist">Enter playlist name</label>
+              <input type="submit" value="Load videos from playlist" /> */}
+            </div>
+          </form>
+          <div id="video-container" />
+        </div>
+      </Grid>
+    )
+  }
 }
