+++ src/defaultData.js
@@ -1,0 +1,124 @@
+export default function defaultData() {
+  return [
+    {
+      title: 'Intermediate moves',
+      tags: ['Marco & Sara', 'Bachatea', '2018'],
+      notes: '',
+      uploadDate: '2019-03-05T10:51',
+      id: '0jP0h3wIZCE',
+      backgroundImageUrl:
+        'http://img.youtube.com/vi/0jP0h3wIZCE/maxresdefault.jpg',
+      status: 0,
+      refresh: false,
+      refreshDate: '',
+    },
+    {
+      title: 'Bachata Flow Freestyle',
+      tags: ['Abdel & Lety', 'Paris Bachata Festival ', '2018', 'Bachata Flow'],
+      notes:
+        'Usher - She came to give it to you (feat. Nicki Minaj) [Dj Bernardo Bachata Remix] ',
+      uploadDate: '2019-03-05T10:51',
+      id: 'VJ0XU8c0yoU',
+      backgroundImageUrl:
+        'http://img.youtube.com/vi/VJ0XU8c0yoU/maxresdefault.jpg',
+      status: 0,
+      refresh: false,
+      refreshDate: '',
+    },
+    {
+      title: 'Body Isolation',
+      tags: ['Marcelo y Belen', 'Europe Bachata Festival', '2018', 'Animals'],
+      notes: '1:30-2:30 favorite part',
+      uploadDate: '2019-03-05T10:51',
+      id: '8GkEhFu7gDs',
+      backgroundImageUrl:
+        'http://img.youtube.com/vi/8GkEhFu7gDs/maxresdefault.jpg',
+      status: 0,
+      refresh: false,
+      refreshDate: '',
+    },
+    {
+      title: 'Bachata Sensual',
+      tags: ['Cornel & Rithika', '2018', 'Head roll'],
+      notes: '',
+      uploadDate: '2019-03-05T10:51',
+      id: 'U-JofUEsbD0',
+      backgroundImageUrl:
+        'http://img.youtube.com/vi/U-JofUEsbD0/maxresdefault.jpg',
+      status: 0,
+      refresh: false,
+    },
+    {
+      title: 'Hip isolation moves',
+      tags: ['Korke y Judith', '2016', 'LA BKS'],
+      notes:
+        'Lorem, ipsum dolor sit amet consectetur dipisicing elit. Voluptates officiis nulla, molestiae tenetur. officiis nulla, molestiae tenetur. offi?',
+      uploadDate: '2019-03-05T10:51',
+      id: 'sGzhU3JJ5Tk',
+      backgroundImageUrl:
+        'http://img.youtube.com/vi/sGzhU3JJ5Tk/maxresdefault.jpg',
+      status: 0,
+      refresh: false,
+    },
+    {
+      title: 'Freestyle moves',
+      tags: ['Lutz & Olivia', '2018', 'Rovinj'],
+      notes:
+        'Lorem, ipsum dolor sit amet consectetur dipisicing elit. Voluptates officiis nulla, molestiae tenetur. officiis nulla, molestiae tenetur. offi?',
+      uploadDate: '2019-03-05T10:51',
+      id: 'b0rovctJe1U',
+      backgroundImageUrl:
+        'http://img.youtube.com/vi/b0rovctJe1U/maxresdefault.jpg',
+      status: 0,
+      refresh: false,
+    },
+    {
+      title: 'Bailemos Despacio',
+      tags: ['Carlos Espinosa y Eleonora', 'Bachata Day', '2019'],
+      notes:
+        'Una Bachata en Boca Chica, Republica Dominicana. Song: Besos a Escondidas Artist: Grupo Extra Ataca y Alemana bailando Bachata de Grupo Extra.',
+      uploadDate: '2019-03-05T10:51',
+      id: 'yeYHmjLijaA',
+      backgroundImageUrl:
+        'http://img.youtube.com/vi/yeYHmjLijaA/maxresdefault.jpg',
+      status: 0,
+      refresh: false,
+    },
+    {
+      title: '10 Intermediate Sensual Figures of Bachata',
+      tags: ['Bachata Day', '2019'],
+      notes:
+        'Una Bachata en Boca Chica, Republica Dominicana. Song: Besos a Escondidas Artist: Grupo Extra Ataca y Alemana bailando Bachata de Grupo Extra.',
+      uploadDate: '2019-03-05T10:51',
+      id: 'tl-gQB6PB-g',
+      backgroundImageUrl:
+        'http://img.youtube.com/vi/tl-gQB6PB-g/maxresdefault.jpg',
+      status: 0,
+      refresh: false,
+    },
+    {
+      title: 'Bachata Sensual Advanced Moves Tutorial',
+      tags: [],
+      notes:
+        'Una Bachata en Boca Chica, Republica Dominicana. Song: Besos a Escondidas Artist: Grupo Extra Ataca y Alemana bailando Bachata de Grupo Extra.',
+      uploadDate: '2019-03-05T10:51',
+      id: 'PPyXdK0PkBU',
+      backgroundImageUrl:
+        'http://img.youtube.com/vi/PPyXdK0PkBU/maxresdefault.jpg',
+      status: 0,
+      refresh: false,
+    },
+    {
+      title: 'Nice wave combination and freestyles',
+      tags: ['Pablo & Raquel', 'Stuttgart', '2018'],
+      notes:
+        'Una Bachata en Boca Chica, Republica Dominicana. Song: Besos a Escondidas Artist: Grupo Extra Ataca y Alemana bailando Bachata de Grupo Extra.',
+      uploadDate: '2019-03-05T10:51',
+      id: 'upTlp-poWoI',
+      backgroundImageUrl:
+        'http://img.youtube.com/vi/upTlp-poWoI/maxresdefault.jpg',
+      status: 0,
+      refresh: false,
+    },
+  ]
+}
